move test factory KubernetesReservationClient to test fixtures module
