+++ subprojects/gradle/instrumentation-tests/src/integTest/kotlin/com/avito/instrumentation/reservation/client/kubernetes/KubernetesReservationClientIntegrationTest.kt
@@ -1,15 +1,10 @@
 package com.avito.instrumentation.reservation.client.kubernetes
 
-import com.avito.instrumentation.reservation.adb.AndroidDebugBridge
-import com.avito.instrumentation.reservation.adb.EmulatorsLogsReporter
 import com.avito.instrumentation.reservation.client.ReservationClient
 import com.avito.instrumentation.reservation.request.Device.CloudEmulator
 import com.avito.instrumentation.reservation.request.Reservation
 import com.avito.instrumentation.reservation.request.createStubInstance
-import com.avito.runner.service.worker.device.adb.Adb
 import com.avito.truth.assertThat
-import com.avito.utils.gradle.KubernetesCredentials
-import com.avito.utils.gradle.createKubernetesClient
 import com.avito.utils.logging.FakeCILogger
 import com.google.common.truth.Truth.assertThat
 import io.fabric8.kubernetes.client.KubernetesClientException
@@ -22,7 +17,6 @@ import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import java.io.File
 import java.net.UnknownHostException
 
 internal class KubernetesReservationClientIntegrationTest {
@@ -34,7 +28,9 @@ internal class KubernetesReservationClientIntegrationTest {
 
     @Test
     fun `claim - throws exception - unknown host`() {
-        clientOne = createClient(
+        clientOne = KubernetesReservationClient.createStubInstance(
+            logger = logger,
+            deploymentNameGenerator = deploymentNameGenerator,
             kubernetesUrl = "unknown-host",
             kubernetesNamespace = "emulators"
         )
@@ -72,8 +68,14 @@ internal class KubernetesReservationClientIntegrationTest {
      */
     @Test
     fun `claim - throws exception - deployment already exists`() {
-        clientOne = createClient()
-        clientTwo = createClient()
+        clientOne = KubernetesReservationClient.createStubInstance(
+            logger = logger,
+            deploymentNameGenerator = deploymentNameGenerator
+        )
+        clientTwo = KubernetesReservationClient.createStubInstance(
+            logger = logger,
+            deploymentNameGenerator = deploymentNameGenerator
+        )
 
         val delayEnoughToCreateFirstDeploymentMs = 3000L
 
@@ -124,50 +126,5 @@ internal class KubernetesReservationClientIntegrationTest {
                 runnable.map { async { it.invoke(this@coroutineScope) } }.awaitAll()
             }
         }
-    }
-
-    private fun createClient(
-        kubernetesUrl: String = requireNotNull(System.getProperty("avito.kubernetes.url")),
-        kubernetesNamespace: String = requireNotNull(System.getProperty("avito.kubernetes.namespace")),
-        configurationName: String = "integration-test",
-        projectName: String = "",
-        buildId: String = "19723577",
-        buildType: String = "",
-        registry: String = ""
-    ): KubernetesReservationClient {
-        val kubernetesCredentials = KubernetesCredentials.Service(
-            token = requireNotNull(System.getProperty("avito.kubernetes.token")),
-            caCertData = requireNotNull(System.getProperty("avito.kubernetes.cert")),
-            url = kubernetesUrl
-        )
-
-        val outputFolder = File("integration")
-        val logcatFolder = File("logcat")
-
-        return KubernetesReservationClient(
-            androidDebugBridge = AndroidDebugBridge(
-                adb = Adb(),
-                logger = { logger.info(it) }
-            ),
-            kubernetesClient = createKubernetesClient(
-                kubernetesCredentials = kubernetesCredentials,
-                namespace = kubernetesNamespace
-            ),
-            emulatorsLogsReporter = EmulatorsLogsReporter(
-                outputFolder = outputFolder,
-                logcatDir = logcatFolder,
-                logcatTags = emptyList()
-            ),
-            logger = logger,
-            reservationDeploymentFactory = ReservationDeploymentFactory(
-                configurationName = configurationName,
-                projectName = projectName,
-                buildId = buildId,
-                buildType = buildType,
-                registry = registry,
-                deploymentNameGenerator = deploymentNameGenerator,
-                logger = logger
-            )
-        )
     }
 }
